+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## master
+### Features / Improvements 🚀
+
+- Support passing an HTMLElement as the parent for a Geocoder - builds on top of [#270] - [#311](https://github.com/mapbox/mapbox-gl-geocoder/pull/311)
 
 ### Bug fixes 🐛
 - disable proximity when reverseGeocoding since it is not supported by the Mapbox Geocoding API [#327](https://github.com/mapbox/mapbox-gl-geocoder/pull/327)